+++ src/components/HoldingsRecordFilters/HoldingsRecordFilters.js
@@ -10,6 +10,9 @@ import {
   CheckboxFilter,
   MultiSelectionFilter,
 } from '@folio/stripes/smart-components';
+import {
+  filterItemsBy,
+} from '../../utils';
 
 import TagsFilter from '../TagsFilter';
 
@@ -75,6 +78,7 @@ export default class HoldingsRecordFilters extends React.Component {
             name="effectiveLocation"
             dataOptions={locationOptions}
             selectedValues={effectiveLocation}
+            filter={filterItemsBy('label')}
             onChange={onChange}
           />
         </Accordion>
@@ -91,6 +95,7 @@ export default class HoldingsRecordFilters extends React.Component {
             name="holdingsPermanentLocation"
             dataOptions={locationOptions}
             selectedValues={holdingsPermanentLocation}
+            filter={filterItemsBy('label')}
             onChange={onChange}
           />
         </Accordion>