+++ src/components/ItemFilters/itemFilterRenderer.test.js
@@ -1,9 +1,13 @@
-import { render, screen } from '@testing-library/react';
+import React from 'react';
+import { screen } from '@testing-library/react';
+import { BrowserRouter as Router } from 'react-router-dom';
 import { noop } from 'lodash';
 
-// import '@folio/stripes-acq-components/test/jest/__mock__';
+import '../../../test/jest/__mock__';
 
 import itemFilterRenderer from './itemFilterRenderer';
+import renderWithIntl from '../../../test/jest/helpers/renderWithIntl';
+import translationsProperties from '../../../test/jest/helpers/translationsProperties';
 
 const DATA = {
   materialTypes: [],
@@ -11,38 +15,35 @@ const DATA = {
   tags: [],
 };
 
-const renderFilters = (data = DATA, onChange = noop) => (render(
-  itemFilterRenderer(data)(onChange)
+const renderFilters = (data = DATA, onChange = noop) => (renderWithIntl(
+  <Router>{itemFilterRenderer(data)(onChange)}</Router>,
+  translationsProperties
 ));
 
-xdescribe('itemFilterRenderer fn', () => {
-  it('displays filter by tags accordion', () => {
-    renderFilters();
-    expect(screen.getByText('ui-inventory.filter.tags')).toBeDefined();
+describe('itemFilterRenderer fn', () => {
+  beforeEach(() => renderFilters());
+
+  it('should display filter by tags accordion', () => {
+    expect(screen.getByText('Tags')).toBeInTheDocument();
   });
 
-  it('displays filter by status accordion', () => {
-    renderFilters();
-    expect(screen.getByText('ui-inventory.item.status')).toBeDefined();
+  it('should display filter by status accordion', () => {
+    expect(screen.getByText('Item status')).toBeInTheDocument();
   });
 
-  it('displays filter by effectiveLocation accordion', () => {
-    renderFilters();
-    expect(screen.getByText('ui-inventory.filters.effectiveLocation')).toBeDefined();
+  it('should display filter by effectiveLocation accordion', () => {
+    expect(screen.getByText('Effective location (item)')).toBeInTheDocument();
   });
 
-  it('displays filter by permanentLocation accordion', () => {
-    renderFilters();
-    expect(screen.getByText('ui-inventory.holdings.permanentLocation')).toBeDefined();
+  it('should display filter by permanentLocation accordion', () => {
+    expect(screen.getByText('Holdings permanent location')).toBeInTheDocument();
   });
 
-  it('displays filter by materialType accordion', () => {
-    renderFilters();
-    expect(screen.getByText('ui-inventory.materialType')).toBeDefined();
+  it('should display filter by materialType accordion', () => {
+    expect(screen.getByText('Material type')).toBeInTheDocument();
   });
 
-  it('displays filter by discoverySuppress accordion', () => {
-    renderFilters();
-    expect(screen.getByText('ui-inventory.discoverySuppress')).toBeDefined();
+  it('should display filter by discoverySuppress accordion', () => {
+    expect(screen.getByText('Suppress from discovery')).toBeInTheDocument();
   });
 });