+++ test/bigtest/interactors/routes/instances-route.js
@@ -4,11 +4,7 @@ import {
   scoped,
 } from '@bigtest/interactor';
 
-import CheckboxInteractor from '@folio/stripes-components/lib/Checkbox/tests/interactor';
-
-import {
-  SearchFieldFilterInteractor,
-} from '../filters';
+import { SearchFieldFilterInteractor } from '../filters';
 
 export default @interactor class InstancesRouteInteractor {
   static defaultScope = '[data-test-inventory-instances]';
@@ -16,6 +12,4 @@ export default @interactor class InstancesRouteInteractor {
   searchFieldFilter = scoped('#pane-filter', SearchFieldFilterInteractor);
   rows = collection('#list-inventory [data-row-index]');
   headers = collection('#list-inventory [data-test-clickable-header]');
-  selectRowCheckboxes = collection('[data-test-inventory-instances] [data-test-select-row]', CheckboxInteractor);
-  customPaneSub = scoped('[data-test-custom-pane-sub]');
 }