+++ test/bigtest/network/config.js
@@ -63,6 +63,7 @@ export default function configure() {
   this.get('/item-damaged-statuses/:id');
 
   this.get('/record-bulk/ids', {});
+  this.post('/data-export/quick-export', {});
 
   this.get('/modes-of-issuance', ({ issuanceModes }) => issuanceModes.all());
   this.get('/modes-of-issuance/:id', ({ issuanceModes }, { params }) => {