+++ test/jest/__mock__/index.js
@@ -1,0 +1,7 @@
+import './currencyData.mock';
+import './documentCreateRange.mock';
+import './matchMedia.mock';
+import './stripesConfig.mock';
+import './stripesCore.mock';
+import './stripesIcon.mock';
+import './stripesSmartComponents.mock';