+++ test/jest/__mock__/stripesSmartComponents.mock.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+jest.mock('@folio/stripes/smart-components', () => ({
+  ...jest.requireActual('@folio/stripes/smart-components'),
+  LocationLookup: () => <div>LocationLookup</div>,
+}), { virtual: true });